Update Dagger to use JDK 18.
This CL sets the default target to JDK 18, which is used by all of the Dagger/Hilt processors. However, we need to be careful with this change to prevent users of Dagger's API from requiring JDK 18. Thus, I've also had to explicitly set all non-processor libraries to target JDK 7 (Dagger) and JDK 8 (Hilt).

In addition, I've added some validation to our GitHub Actions to check that the classes in our artifact jars all have the expected language level.

RELNOTES=N/A
PiperOrigin-RevId: 654943496